+++ src/index.ts
@@ -1,8 +1,10 @@
 export * from "./byte-queue.js";
 export * from "./bytes.js";
 export * from "./deserializer.js";
+export * from "./encoding/base32.js";
 export * from "./encoding/base64.js";
 export * from "./encoding/hex.js";
 export * from "./endianess.js";
 export * from "./serializer.js";
+export { setupSharedBuffer } from "./shared-buffers.js";
 export * from "./stream-deserializer.js";